Fix CRC input type: cast frame bit vectors to double
comm.CRCGenerator/CRCDetector require double or logical input. de2bi can
return uint8, which promoted the concatenated frame bit vector to uint8 and
made CRCGenerator error out in build_frames. Cast the payload bits and the
protected/CRC bit vectors to double on both the transmit and receive sides.

+++ run_receiver.m
@@ -144,7 +144,7 @@ crcBits       = bits(48 + 8*len + 1 : need);
 protectedBits = bits(1 : 48 + 8*len);
 
 % CRC-16 check over [sync|seq|len|payload] against the received CRC.
-[~, err] = crcDet([protectedBits; crcBits]);
+[~, err] = crcDet(double([protectedBits; crcBits]));
 frame.crcValid = (err == 0);
 
 frame.payloadLen = len;
